Patch Chef internals for mutating array operations
There is a bug in Chef, preventing detection of mutating array ops.
See https://github.com/chef/chef/issues/13995

However, the fix is not yet merged, and even when it will be it may
not be available for people not able to bump Chef.

This commit patches Chef internales to backport the fix provided in:
https://github.com/chef/chef/issues/13996

This should fix #7.