CI: Test only on Ubuntu but multiple ruby versions (2.7.8, 3.1.4)
While the cookbook aim to be compatible with older Chef version, it is
important to keep running a not to old Ruby in the CI.
So use 2.7.8 for legacy ruby, and also add the latest 3.x compatible.
Note that Chef < 18 is not compatible with ruby >= 3.2.

Also remove MacOS & Windows from the test matrix, the cookbook is not
platform dependent. No need to use extra resources for nothing.